requirements.txt
asn1crypto==0.24.0
astroid==2.1.0
cffi==1.11.5
colorama==0.4.1
cryptography==2.5
Django==2.1.7
django-rest-knox==3.6.0
djangorestframework==3.9.1
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pycparser==2.19
pylint==2.2.2
pyOpenSSL==19.0.0
pytz==2018.9
six==1.12.0
typed-ast==1.3.1
wrapt==1.11.1
